Refactor stdcommands into command packages.
There is no new functionality here. This is merely a step toward Go ideals of
accept interfaces return structs.

The intent is to simplify dependencies, make it easier to test, and make
command additions almost as simple as coping a directory and modifying. I kept
the manual command list in stdcommands to preserve existing command ordering in
the help, but a future implementation may simply use a sorted, implicit
registry.
