+++ stdcommands/catfact/command.go
@@ -1,0 +1,32 @@
+package catfact
+
+import (
+	"github.com/jonas747/dcmd"
+	"github.com/jonas747/yagpdb/commands"
+	"github.com/jonas747/yagpdb/stdcommands/util"
+	"math/rand"
+)
+
+var yagCommand = commands.YAGCommand{
+	CmdCategory: commands.CategoryFun,
+	Name:        "CatFact",
+	Aliases:     []string{"cf", "cat", "catfacts"},
+	Description: "Cat Facts",
+
+	RunFunc: func(data *dcmd.Data) (interface{}, error) {
+		cf := Catfacts[rand.Intn(len(Catfacts))]
+		return cf, nil
+	},
+}
+
+func Cmd() util.Command {
+	return &cmd{}
+}
+
+type cmd struct {
+	util.BaseCmd
+}
+
+func (c cmd) YAGCommand() *commands.YAGCommand {
+	return &yagCommand
+}
