+++ stdcommands/leaveserver/command.go
@@ -1,0 +1,40 @@
+package leaveserver
+
+import (
+	"github.com/jonas747/dcmd"
+	"github.com/jonas747/yagpdb/commands"
+	"github.com/jonas747/yagpdb/common"
+	"github.com/jonas747/yagpdb/stdcommands/util"
+)
+
+var yagCommand = commands.YAGCommand{
+	Cooldown:             2,
+	CmdCategory:          commands.CategoryDebug,
+	HideFromCommandsPage: true,
+	Name:                 "leaveserver",
+	Description:          ";))",
+	HideFromHelp:         true,
+	RequiredArgs:         1,
+	Arguments: []*dcmd.ArgDef{
+		{Name: "server", Type: dcmd.Int},
+	},
+	RunFunc: util.RequireOwner(func(data *dcmd.Data) (interface{}, error) {
+		err := common.BotSession.GuildLeave(data.Args[0].Int64())
+		if err == nil {
+			return "Left " + data.Args[0].Str(), nil
+		}
+		return err, err
+	}),
+}
+
+func Cmd() util.Command {
+	return &cmd{}
+}
+
+type cmd struct {
+	util.BaseCmd
+}
+
+func (c cmd) YAGCommand() *commands.YAGCommand {
+	return &yagCommand
+}
