+++ stdcommands/reverse/command.go
@@ -1,0 +1,41 @@
+package reverse
+
+import (
+	"github.com/jonas747/dcmd"
+	"github.com/jonas747/yagpdb/commands"
+	"github.com/jonas747/yagpdb/stdcommands/util"
+)
+
+var yagCommand = commands.YAGCommand{
+	CmdCategory:  commands.CategoryFun,
+	Name:         "Reverse",
+	Aliases:      []string{"r", "rev"},
+	Description:  "Reverses the text given",
+	RunInDM:      true,
+	RequiredArgs: 1,
+	Arguments: []*dcmd.ArgDef{
+		&dcmd.ArgDef{Name: "What", Type: dcmd.String},
+	},
+	RunFunc: func(data *dcmd.Data) (interface{}, error) {
+		toFlip := data.Args[0].Str()
+
+		out := ""
+		for _, r := range toFlip {
+			out = string(r) + out
+		}
+
+		return ":upside_down: " + out, nil
+	},
+}
+
+func Cmd() util.Command {
+	return &cmd{}
+}
+
+type cmd struct {
+	util.BaseCmd
+}
+
+func (c cmd) YAGCommand() *commands.YAGCommand {
+	return &yagCommand
+}
