+++ stdcommands/secretcommand/command.go
@@ -1,0 +1,33 @@
+package secretcommand
+
+import (
+	"github.com/jonas747/dcmd"
+	"github.com/jonas747/discordgo"
+	"github.com/jonas747/yagpdb/commands"
+	"github.com/jonas747/yagpdb/common"
+	"github.com/jonas747/yagpdb/stdcommands/util"
+)
+
+var yagCommand = commands.YAGCommand{
+	Cooldown:             2,
+	CmdCategory:          commands.CategoryDebug,
+	HideFromCommandsPage: true,
+	Name:                 "secretcommand",
+	Description:          ";))",
+	HideFromHelp:         true,
+	RunFunc: util.RequireOwner(func(data *dcmd.Data) (interface{}, error) {
+		return "<@" + discordgo.StrID(common.Conf.Owner) + "> Is my owner", nil
+	}),
+}
+
+func Cmd() util.Command {
+	return &cmd{}
+}
+
+type cmd struct {
+	util.BaseCmd
+}
+
+func (c cmd) YAGCommand() *commands.YAGCommand {
+	return &yagCommand
+}
